[refactor] 상위 폴더 application 생성 및 모듈 구조 리팩터링
