+++ src/application/auth/auth.controller.ts
@@ -1,0 +1,21 @@
+import { Controller, Post, Body, HttpCode, HttpStatus } from '@nestjs/common';
+import { AuthService } from './auth.service';
+import { LoginDto } from 'src/application/DB/users/dto/login.dto';
+import { CreateUserDto } from 'src/application/DB/users/dto/create-user.dto';
+import { User } from 'src/application/DB/users/schemas/user.schema';
+
+@Controller('auth')
+export class AuthController {
+  constructor(private authService: AuthService) {}
+
+  @Post('/register')
+  async register(@Body() createUserDto: CreateUserDto): Promise<User> {
+    return this.authService.registerUser(createUserDto);
+  }
+
+  @Post('/login')
+  @HttpCode(HttpStatus.OK)
+  async login(@Body() loginDto: LoginDto) {
+    return this.authService.login(loginDto);
+  }
+}
