+++ src/application/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { UsersModule } from 'src/users/users.module';
+import { UsersModule } from 'src/application/DB/users/users.module';
 import { JwtModule } from '@nestjs/jwt';
 
 @Module({
